actions moved inside of app
